pyproject.toml
[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[project]
name = "ij"
version = "0.2.0"
description = "Idea Junction - Bidirectional diagramming system with AI, code analysis, and multi-format support"
readme = "README.md"
requires-python = ">=3.10"
keywords = ["diagrams", "flowcharts", "visualization", "mermaid", "d2", "plantuml", "ai"]
authors = []
dependencies = [
    "networkx>=3.0",
    "graphviz>=0.20",
]

[project.license]
text = "mit"

[project.urls]
Homepage = "https://github.com/i2mint/ij"
Repository = "https://github.com/i2mint/ij"

[project.scripts]
ij = "ij.cli:main"
ij2 = "ij.cli_enhanced:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
]
docs = [
    "sphinx>=6.0",
    "sphinx-rtd-theme>=1.0",
]
ai = [
    "openai>=1.0.0",
]
export = [
    "playwright>=1.40.0",
]
database = [
    "sqlalchemy>=2.0.0",
]
full = [
    "openai>=1.0.0",
    "playwright>=1.40.0",
    "sqlalchemy>=2.0.0",
    "pyyaml>=6.0",
]

[tool.ruff]
line-length = 88
target-version = "py310"
exclude = [
    "**/*.ipynb",
    ".git",
    ".venv",
    "build",
    "dist",
    "tests",
    "examples",
    "scrap",
]

[tool.ruff.lint]
select = [
    "D100",
]
ignore = [
    "D203",
    "E501",
    "B905",
]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "D",
]
"examples/*" = [
    "D",
]
"scrap/*" = [
    "D",
]

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [
    "tests",
]
doctest_optionflags = [
    "NORMALIZE_WHITESPACE",
    "ELLIPSIS",
]
